Update README.md
docs: Update README with information about The Mandalorian

Added details about the TV series "The Mandalorian" to the README file. Included a brief synopsis, key elements, memorable quotes, and a personal note on why I love the series. Formatted the content to be more readable in the context of a README file.